contenu des here-doc récupérés, et écrasés si nécessaire
