+++ exec/exec.c
@@ -203,152 +203,6 @@ void	pipex_loop(t_msh *msh)
 	}
 }
 
-void	add_hd_node(t_msh *msh, char *line)
-{
-	t_hd	*hd_node;
-	t_hd	*new;
-
-	hd_node = msh->exec->hd;
-	new = new_hd();
-	if (new)
-	{
-		while (hd_node->next)
-			hd_node = hd_node->next;
-	}
-	hd_node->next = new;
-	new->str = ft_strdup(line);
-}
-/*
-void	get_here_doc_content(t_ppx *ppx)
-{
-	char	*line;
-	t_hdc	*node;
-
-	line = NULL;
-	node = NULL;
-	ppx->delimiter = ft_strjoin(ppx, ppx->av[2], "\n");
-	while (1)
-	{
-		write(STDIN_FILENO, "> ", 2);
-		line = get_next_line(ppx, STDIN_FILENO);
-		if (ft_strcmp(line, ppx->delimiter))
-			break ;
-		node = (t_hdc *)malloc(sizeof(t_hdc));
-		if (!node)
-			exit(EXIT_FAILURE);
-		node->s = ft_strdup(ppx, line);
-		append_node_hdcstring(ppx, node);
-		free_chars(&line);
-	}
-	free_chars(&line);
-}
-*/
-
-t_bool hd_strcmp(char *limiter, char *line)
-{
-	char	*limiter_with_nl;
-	t_bool	return_value;
-
-	limiter_with_nl = ft_strjoin(limiter, "\n");
-	if (ft_strcmp(limiter_with_nl, line))
-		return_value = TRUE;
-	else
-		return_value = FALSE;
-	free(limiter_with_nl);
-	return (return_value);
-}
-
-void	get_hd(t_exec_list *exec_list_node, int i)
-{
-	char	*line;
-	t_hd	*hd_node; // continuer ici
-
-	line = NULL;
-	while (42)
-	{
-		write(STDOUT_FILENO, "> ", 2);
-		line = get_next_line(STDIN_FILENO);
-		if (hd_strcmp(exec_list_node->redirect_array[i].str, line))
-			break ;
-		
-
-
-		else
-			msh->exec->hd->str = ft_strdup(line);
-		free(line);
-		i++;
-	}
-	free(line);
-}
-
-void	clear_hd(t_exec_list *exec_list_node)
-{
-	t_hd	*hd_node;
-	t_hd	*hd_node_next;
-
-	hd_node = exec_list_node->hd;
-	hd_node_next = NULL;
-	while (hd_node)
-	{
-		hd_node_next = hd_node->next;
-		if (hd_node->str)
-		{
-			free(hd_node->str);
-			hd_node->str = NULL;
-		}
-		free(hd_node);
-		hd_node = hd_node_next;
-	}
-}
-
-void	get_hd_in_exec_list_node(t_exec_list *exec_list_node)
-{
-	int	i;
-
-	i = -1;
-	while (++i < exec_list_node->nb_redirects)
-	{
-		if (exec_list_node->redirect_array[i].exp_type == LIMITER_HEREDOC)
-		{
-			if (exec_list_node->hd->str)
-			{
-				clear_hd(exec_list_node);
-				exec_list_node->hd = new_hd();
-			}
-			get_hd(exec_list_node, i);
-		}
-	}
-}
-
-t_bool	hd_in_exec_list_node(t_exec_list *exec_list_node)
-{
-	int		i;
-
-	i = -1;
-	while (++i < exec_list_node->nb_redirects)
-	{
-		if (exec_list_node->redirect_array[i].exp_type == LIMITER_HEREDOC)
-			return (TRUE);
-	}
-	return (FALSE);
-}
-
-void	get_all_hd_content(t_msh *msh)
-{
-	t_exec_list	*exec_list_node;
-
-	exec_list_node = msh->exec_list;
-	while (exec_list_node)
-	{
-		if (hd_in_exec_list_node(exec_list_node))
-		{
-			exec_list_node->hd = new_hd();
-			get_hd_in_exec_list_node(exec_list_node);
-		}
-		exec_list_node = exec_list_node->next;
-	}
-}
-
 void execution(t_msh *msh, int ac, char **av, char **envp)
 {
 	msh->exec = new_exec();
@@ -357,6 +211,8 @@ void execution(t_msh *msh, int ac, char **av, char **envp)
 	feed_msh_acavenvp(msh, ac, av, envp);
 
 	get_all_hd_content(msh);
+
+	print_all_hd_remaining(msh);
 
 	//print_paths_from_path(msh);
