+++ include/fn_dev.h
@@ -57,6 +57,16 @@ char	*gnl_strdup(char *s);
 char	*gnl_strjoin(char *s1, char *s2);
 char	*gnl_substr(char *s, unsigned int start, size_t len);
 
+
+t_bool	hd_strcmp(char *limiter, char *line);
+void	feed_append_new_hd_node(t_exec_list *exec_list_node, char *line);
+void	get_hd(t_exec_list *exec_list_node, int i);
+void	clear_hd(t_exec_list *exec_list_node);
+void	get_hd_in_exec_list_node(t_exec_list *exec_list_node);
+t_bool	hd_in_exec_list_node(t_exec_list *exec_list_node);
+void	get_all_hd_content(t_msh *msh);
+void	print_all_hd_remaining(t_msh *msh);
+
 t_exec	*new_exec(void);
 t_hd	*new_hd(void);
 void	execution(t_msh *msh, int ac, char **av, char **envp);
