+++ utils/ft_strcmp.c
@@ -9,7 +9,7 @@ int	ft_strcmp(char *s1, char *s2)
 		return (0);
 	if (ft_strlen(s1) != ft_strlen(s2))
 		return (0);
-	while (s1[i] == s2[i])
+	while (s1[i] && s1[i] == s2[i])
 		i++;
 	if (s1[i] || s2[i])
 		return (0);
